Started reworking internal calculator functionality. This commit introduces serious regressions in functionality. Up until now, the calculator would recalculate every line in the window every time a key was pressed by the user. Clearly this is inefficient and unnecessary. I reached a level of functionality that I am happy with for an initial release, so now begins the task of making Spark Calc work efficiently.
The new internal design for the calculator uses what I'm going to call a Sheet. Each sheet stores the string lines that the user has typed into the input box, along with the numerical value of each line, and the variable name associated with that line if applicable. Lines that represent variables will store a set of dependent lines that will also need to be recalculated, in the event the numerical value of that line changes. This ultimately means that only the lines affected by a change the user makes are recalculated. Separating this functionality into sheets also makes it easier to add multiple pages for input into the GUI later on.